Aviation_Maintenance_App/requirements.txt
flask
pandas==2.0.3
gunicorn
supabase
google-genai
python-dotenv
httpx>=0.24,<0.28